Update features_extraction.py
Created a function which evaluates the plane at a certain x and y point based off the equation of the plane (didn't know how to create an equation of a plane like how you can in Mathematica). Create most of the section relating to filtering patches at waist, neck, and shoulders. Did not know how to proceed with parts commented as "#Add action" based off the similar sequencing in the Mathematica code.